+++ src/Batch/AccountBalanceBatch.php
@@ -46,6 +46,7 @@ class AccountBalanceBatch extends AbstractBatch
 
     public function getBalance(): AccountBalance
     {
+        $this->requireValidPosition('read balance');
         $offset = $this->currentPosition * $this->getStructSize();
         $data = \substr($this->buffer, $offset, $this->getStructSize());
         $unpacked = BinaryHelper::unpackAccountBalance($data);